Added charge sharing, simple model.
The simple model of charge charing does the following:
If the particle hits the sensor in the charge sharing range 		     		(if (Rndm() < csRange / pitch )) {
then the charge is shared linearly between strips, depending on its location, x		(x = Rndm(); )
									     		(signal *= 0.5 + x/2; )

the 0.5 is the offset, if the charge is exactly in the center, both strips get half of the charge.
x is devided by 2 because we simulate only one strip, and we put x exactly between two strips. So instead of x going between 0 and 1, it should go between 0 and 0.5.